requirements.txt
alembic==1.6.2
azure-ai-textanalytics==5.0.0
azure-cognitiveservices-vision-face==0.5.0
azure-common==1.1.27
azure-core==1.14.0
azure-nspkg==3.0.2
azure-storage-blob==12.8.1
azure-storage-common==2.1.0
Babel==2.9.1
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.4
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
flask-googlemaps==0.4.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
google-api-core==1.26.3
google-auth==1.30.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
htmlmin==0.1.12
idna==2.10
importlib-metadata==4.0.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsmin==2.2.2
livereload==2.6.3
login==0.0.6
lunr==0.5.8
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.0
mkdocs==1.1.2
mkdocs-material==7.1.4
mkdocs-material-extensions==1.0.1
mkdocs-minify-plugin==0.4.0
msrest==0.6.21
nltk==3.6.2
oauthlib==3.1.0
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
opencensus-ext-flask==0.7.5
packaging==20.9
protobuf==3.17.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pymdown-extensions==8.2
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
tornado==6.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
